+++ docs/workshop/researcher/researcher-1.prompty
