+++ docs/workshop/socialmedia/social.prompty
@@ -10,7 +10,7 @@ model:
     azure_deployment: gpt-4
     api_version: 2023-07-01-preview
   parameters:
-    max_tokens: 1500
+    max_tokens: 100
 sample:
   researchContext: Can you find the latest camping trends and what folks are doing in the winter?
   research: research
